Send color frames from server to netclient
The server was obstructed from sending frames
because its own usermeshRenderer initialized
the pointcloudcolortexture which was required
to be null to initialize the field responsible
for sending the data. This dependency caused issues
and adding a manner in which both can be initialized
open the doors to the frames being sent.
Ba-da-bing, ba-da-boom!